Move send reset email to own module
